requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
python-dotenv==1.0.1
django-environ==0.11.2
django-filter==23.5
Pillow==10.2.0
requests==2.31.0
openai>=1.0.0
anthropic>=0.7.0
pdf2image>=1.17.0
boto3>=1.34.0
textract-trp>=0.1.3
amazon-textract-caller>=0.2.0
amazon-textract-prettyprinter>=0.1.10
amazon-textract-response-parser>=0.1.39,<0.2 